Add show, preview, and download for file

+++ resources/views/livewire/archive-box/file/index.blade.php
@@ -8,8 +8,7 @@
     <x-card>
         <x-table :headers="$this->headers" :rows="$files" :sort-by="$sortBy" with-pagination >
             @scope('cell_slug', $file)
-                {{-- maybe view file ? --}}
-                <x-button label="{{ $file->slug }}" link="#" icon="o-document" tooltip="{{ 'View '.$file->name }}" responsive class="w-fit flex-nowrap" />
+                <x-button label="{{ $file->slug }}" link="{{ route('file.show', $file->slug) }}" icon="o-document" tooltip="{{ 'View '.$file->name }}" responsive class="w-fit flex-nowrap" />
             @endscope
             @scope('cell_extension', $file)
                 <x-badge value="{{ $file->extension }}" class="badge-info whitespace-nowrap" />
@@ -17,16 +16,33 @@
             @scope('cell_size', $file)
                 <x-badge value="{{ Number::fileSize($file->size) }}" class="badge-info whitespace-nowrap" />
             @endscope
-            @scope('cell_actions', $file)
+            @scope('cell_actions', $file, $archiveBox)
                 <div class="w-fit h-fit flex flex-row space-x-2 space-y-0">
-                    <x-button label="Edit" wire:click="editFile({{ $file->id }})" icon="o-pencil" tooltip="{{ 'Edit '.$file->name }}" responsive class="w-fit flex-nowrap" />
-                    <x-button label="Delete" wire:click="deleteFile({{ $file->id }})" wire:confirm="{{ 'Delete '.$file->name.' ?' }}" icon="o-trash" tooltip="{{ 'Delete '.$file->name }}" responsive class="w-fit flex-nowrap" />
+                    @auth
+                        @if ($archiveBox->users->whereIn('pivot.permission', [2,3])->contains(Auth::id()))
+                            <x-button label="Edit" wire:click="editFile({{ $file->id }})" icon="o-pencil" tooltip="{{ 'Edit '.$file->name }}" responsive class="w-fit flex-nowrap" />
+                            @if ($archiveBox->users->where('pivot.permission', 3)->contains(Auth::id()))
+                                <x-button label="Delete" wire:click="deleteFile({{ $file->id }})" wire:confirm="{{ 'Delete '.$file->name.' ?' }}" icon="o-trash" tooltip="{{ 'Delete '.$file->name }}" responsive class="w-fit flex-nowrap" />
+                            @endif
+                        @endif
+                    @endauth
+                    @switch($archiveBox->private)
+                        @case(false)
+                            <x-button label="Download" link="{{ route('file.download', $file->slug) }}" icon="o-document-arrow-down" tooltip="{{ 'Download '.$file->name }}" no-wire-navigate responsive class="w-fit flex-nowrap" />
+                            @break
+                        @default
+                            @auth
+                                @if ($archiveBox->users->whereIn('pivot.permission', [1,2,3])->contains(Auth::id()))
+                                    <x-button label="Download" link="{{ route('file.download', $file->slug) }}" icon="o-document-arrow-down" tooltip="{{ 'Download '.$file->name }}" no-wire-navigate responsive class="w-fit flex-nowrap" />
+                                @endif
+                            @endauth
+                    @endswitch
                 </div>
             @endscope
         </x-table>
     </x-card>
     @auth
-        @if ($archiveBox->users->where('pivot.permission', 3)->contains(Auth::id()))
+        @if ($archiveBox->users->whereIn('pivot.permission', [2,3])->contains(Auth::id()))
             <x-drawer
                 wire:model="showEditFile"
                 title="Edit File"
@@ -34,7 +50,7 @@
                 separator
                 with-close-button
                 close-on-escape
-                class="w-11/12 lg:w-1/3">
+                class="w-fit">
                 @livewire('archive-box.file.edit', ['archiveBox' => $archiveBox])
             </x-drawer>
         @endif
